% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyMS_R6_TransitionCorrelations.R
\name{aggregateTopNIntensities}
\alias{aggregateTopNIntensities}
\title{aggregates top N intensities}
\usage{
aggregateTopNIntensities(data, config, N = 3)
}
\description{
run \link{rankPrecursorsByIntensity} first
}
\examples{
library(SRMService)
config <- spectronautDIAData250_config$clone(deep=T)
res <- setLarge_Q_ValuesToNA(spectronautDIAData250_analysis, config)
res <- rankPrecursorsByIntensity(res,config)
aggregateTopNIntensities(res, config, N=3)
}
