% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_ProteinTable.R
\docType{class}
\name{Annotation}
\alias{Annotation}
\title{Condition Map}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Annotation
}
\description{
Condition Map
}
\examples{

library(SRMService)
x <- data.frame(Raw.file = c("A","B","C"),"MeasurementOrder" = c(1,2,3) , Gender = c("F","F","M"))
cm <- Annotation$new(x, experimentID = "E1 with random Effect", fixed = "Gender", random="MeasurementOrder")
cm$exists("Gender")
cm$exists("Stop")
cm$fixed
cm$random
cm$as_numeric("Gender")
cm$levels("Gender")
cm$get("Gender")
cm$get_color("Gender")
cm$sampleID
cm$get_sample_id()
cm$annotation
}
\keyword{datasets}
