% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RequiredColumns.R
\docType{data}
\name{annotationColumns}
\alias{annotationColumns}
\title{Annotation Columns}
\format{
An object of class \code{character} of length 5.
}
\usage{
annotationColumns
}
\description{
Annotation Columns
}
\keyword{datasets}
