% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyMS_R6_TransitionCorrelations.R
\name{applyMatrixFunction}
\alias{applyMatrixFunction}
\title{apply Function To matrix}
\usage{
applyMatrixFunction(data, config, normalization)
}
\description{
apply Function To matrix
}
\examples{

conf <- skylineconfig$clone(deep = TRUE)
res <- applyMatrixFunction(sample_analysis, conf, normalization = base::scale)
stopifnot("Area_base..scale" \%in\% colnames(res))
stopifnot("Area_base..scale" == conf$table$getWorkIntensity())

res <- applyMatrixFunction(res, conf, normalization = robust_scale)
}
