% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyMS_R6_AnalysisConfiguration.R
\name{applyToHierarchyBySample}
\alias{applyToHierarchyBySample}
\title{applys func - a funciton workin on matrix for each protein and returning a vector of the same length as the number of samples}
\usage{
applyToHierarchyBySample(data, config, func, hierarchy_level = 1,
  unnest = FALSE)
}
\description{
applys func - a funciton workin on matrix for each protein and returning a vector of the same length as the number of samples
}
\examples{
x <- applyToHierarchyBySample(sample_analysis, skylineconfig, medpolishPly)

x \%>\% dplyr::select(skylineconfig$table$hierarchyKeys()[1] ,  medpolishPly) \%>\% unnest()

}
