% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyMS_R6_AnalysisConfiguration.R
\name{applyToHighestHierarchyBySample}
\alias{applyToHighestHierarchyBySample}
\title{applys func - a funciton workin on matrix for each protein and returning a vector of the same length as the number of samples}
\usage{
applyToHighestHierarchyBySample(data, config, func)
}
\description{
applys func - a funciton workin on matrix for each protein and returning a vector of the same length as the number of samples
}
\examples{
x <- applyToHighestHierarchyBySample(sample_analysis, skylineconfig, medpolishPly)

x \%>\% dplyr::select(config$table$hierarchyKeys()[1] ,  medpolishPly) \%>\% unnest()

}
