% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhosphoService_HelperFunctions_v4_2019.R
\name{buildUpAnnotationFileFromColumnNames}
\alias{buildUpAnnotationFileFromColumnNames}
\title{This function builds up the Annotations from the raw-file names if they follow the naming conventions}
\usage{
buildUpAnnotationFileFromColumnNames(proteinMAT,
  numericSeperatorForFilenames = 5)
}
\arguments{
\item{proteinMAT}{proteinMatrix object as read in}

\item{numericSeperatorForFilenames}{id_x where the conditions are encoded in the file names}
}
\value{
annotation table
}
\description{
This function builds up the Annotations from the raw-file names if they follow the naming conventions
}
