% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhosphoService_HelperFunctions_v4_2019.R,
%   R/PhosphoService_HelperFunctions_v5_2020.R
\name{combineAndReturnPhosphoMatrix}
\alias{combineAndReturnPhosphoMatrix}
\title{Simply merge phosphoAnalysis with total proteinAnalysis}
\usage{
combineAndReturnPhosphoMatrix(mergedResultMatrix, globalProteinQuantResults)

combineAndReturnPhosphoMatrix(mergedResultMatrix, globalProteinQuantResults)
}
\arguments{
\item{mergedResultMatrix}{result matrix with statistics from 2grp analysis and more information from PhosphoSTY}

\item{globalProteinQuantResults}{result matrix with statistics from 2grp analysis for total protein}
}
\value{
merged result lists with doublications on protein side (y)

merged result lists with doublications on protein side (y)
}
\description{
Simply merge phosphoAnalysis with total proteinAnalysis

Simply merge phosphoAnalysis with total proteinAnalysis
}
\examples{
combineAndReturnPhosphoMatrix(mergedResultMatrix = SRMService::res__merge_allPhosPep, globalProteinQuantResults = SRMService::quantResTotalProt)

combineAndReturnPhosphoMatrix(mergedResultMatrix = SRMService::res__merge_allPhosPep, globalProteinQuantResults = SRMService::quantResTotalProt)

}
