% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyMS_R6_Modelling.R
\name{compute_roc}
\alias{compute_roc}
\title{Apply roc analysis on main factor on lowest level}
\usage{
compute_roc(data, config)
}
\description{
Apply roc analysis on main factor on lowest level
}
\examples{
library(tidyverse)
library(SRMService)
config <- spectronautDIAData250_config$clone(deep=T)
config$parameter$min_nr_of_notNA  <- 20
data <- spectronautDIAData250_analysis
x <- sample(data$protein_Id,2)
data <- data \%>\% filter(protein_Id \%in\% x)
res <- compute_roc(data, config)
unique(res$protein_Id)
}
