% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_AnalysisConfiguration.R
\name{configurationPlotPeptides}
\alias{configurationPlotPeptides}
\title{extracts the relevant information from the configuration to make the plot.}
\usage{
configurationPlotPeptides(res, proteinName, configuration, separate = FALSE)
}
\description{
extracts the relevant information from the configuration to make the plot.
}
