% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limmaHelpers.R
\name{contrasts.fit.NA}
\alias{contrasts.fit.NA}
\title{version of limma::contrast.fit which is able to handle NA's in one of the conditions.}
\usage{
contrasts.fit.NA(fit, cont)
}
\arguments{
\item{fit}{returned by lmFit}

\item{cont}{returned by limma::makeContrasts}
}
\value{
tibble with pValues and fold changes
}
\description{
version of limma::contrast.fit which is able to handle NA's in one of the conditions.
}
\examples{

\dontrun{
if(FALSE){
cont <- limma::makeContrasts(contrasts = contrasts, levels =  levels)
fit <- lmFit(intmat , designMatrix)
res <- contrasts.fit.NA(fit,cont)
}
}
}
