% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyMS_R6_AnalysisConfiguration.R
\name{craeteSkylineConfiguration}
\alias{craeteSkylineConfiguration}
\title{This function sets up an example configuration}
\usage{
craeteSkylineConfiguration(isotopeLabel = "Isotope.Label",
  qValue = "annotation_QValue")
}
\description{
This function sets up an example configuration
}
