% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhosphoService_HelperFunctions_v4_2019.R
\name{doSequenceLogoPlotsFromCandidateMatrixForSinglePhosPeptides}
\alias{doSequenceLogoPlotsFromCandidateMatrixForSinglePhosPeptides}
\title{This functions generates a sequence logo from sequence windows (usually 31 chars long, PTMsite in the middle)}
\usage{

  doSequenceLogoPlotsFromCandidateMatrixForSinglePhosPeptides(merged_phosphoPeptideResultMatrix,
  phosAcceptorAA = "S", myLogoPlotTitle = "Sequence Logo")
}
\arguments{
\item{merged_phosphoPeptideResultMatrix}{combinded phospho result matrix with stats and info from PhosphoSTY}

\item{phosAcceptorAA}{Phospho acceptor site (in case of phospho: S|T|Y)}

\item{myLogoPlotTitle}{Title for the plot}
}
\description{
This functions generates a sequence logo from sequence windows (usually 31 chars long, PTMsite in the middle)
}
