% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eb.fit.R
\name{eb.fit}
\alias{eb.fit}
\title{fit limma}
\usage{
eb.fit(dat, design)
}
\arguments{
\item{dat}{data matrix}

\item{design}{design matrix}
}
\description{
fit limma
}
\examples{
data <- matrix(rnorm(120),ncol=6)
x <- c(rep("a",3),rep("b",3))
x <- model.matrix(~x)
eb.fit(data,x)

}

