% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyMS_R6_AnalysisConfiguration.R
\name{extractIntensities}
\alias{extractIntensities}
\title{Extract intensity column}
\usage{
extractIntensities(x, configuration)
}
\description{
Extract intensity column
}
\examples{
library(dplyr)
xnested <- sample_analysis \%>\%
 group_by_at(skylineconfig$table$hierarchyKeys()[1]) \%>\%
 tidyr::nest()
xx <- extractIntensities(xnested$data[[1]],skylineconfig)
stopifnot(dim(xx)==c(104,22))
}
