% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/One2OneAnalysisHelpers.R
\name{fgcz_render_One2OneReport}
\alias{fgcz_render_One2OneReport}
\title{perform rendering}
\usage{
fgcz_render_One2OneReport(maxquanttxtdirectory = ".",
  reportFileBaseName = "fgcz_MQ_QC_report")
}
\description{
perform rendering
}
