% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDProteins.R
\name{fix_PD_inputFiles}
\alias{fix_PD_inputFiles}
\title{massage PD inputFiles.txt - remove columns and fix column names}
\usage{
fix_PD_inputFiles(inputFiles)
}
\arguments{
\item{inputFiles}{read _InputFiles.txt file using the readr::read_tsv and pass it to inputFiles}
}
\description{
massage PD inputFiles.txt - remove columns and fix column names
}
