% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScaffoldHelpers.R
\name{fixScaffoldQuantReport}
\alias{fixScaffoldQuantReport}
\title{fix scaffold file (utility function)}
\usage{
fixScaffoldQuantReport(filename)
}
\arguments{
\item{filename}{filename of scaffod quant reprot}
}
\description{
fix scaffold file (utility function)
}

