% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyMS_R6_TransitionCorrelations.R
\name{gatherItBack}
\alias{gatherItBack}
\title{make it long}
\usage{
gatherItBack(x, value, config, data = NULL)
}
\description{
make it long
}
\examples{
res <- toWideConfig(sample_analysis, skylineconfig, as.matrix = TRUE)
res <- scale(res)
xx <- gatherItBack(res,"srm_intensityScaled",skylineconfig)
xx <- gatherItBack(res,"srm_intensityScaled",skylineconfig,sample_analysis)
}
