% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhosphoService_HelperFunctions_v4_2019.R
\name{generateNtoCProteinPDFsWithPhosphoPeptides_2019}
\alias{generateNtoCProteinPDFsWithPhosphoPeptides_2019}
\title{Generate PDFs with NtoCplots for all candidates (using all phospho peptides even if not significant)}
\usage{

  generateNtoCProteinPDFsWithPhosphoPeptides_2019(globalNphosphoCombinedNResultMatrix,
  candidateMatrix, expName)
}
\arguments{
\item{globalNphosphoCombinedNResultMatrix}{combinded result matrix with phospho and total protein result}

\item{candidateMatrix}{contains the significant phospho peptides for which all proteins should be drawn}

\item{expName}{this name is used to label the pdf file that is generated with this function}
}
\description{
Generate PDFs with NtoCplots for all candidates (using all phospho peptides even if not significant)
}
