% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhosphoService_HelperFunctions_v4_2019.R,
%   R/PhosphoService_HelperFunctions_v5_2020.R
\name{generateTxtFilesForDownstreamAnalysisAndReturnCandidateMatrix_2019}
\alias{generateTxtFilesForDownstreamAnalysisAndReturnCandidateMatrix_2019}
\title{Write txt files with unique protein names that can be used for ORA analysis}
\usage{
generateTxtFilesForDownstreamAnalysisAndReturnCandidateMatrix_2019(
  comparisonName = "TestComparison",
  mergedResultMatrix,
  my_pModThreshold = pModThreshold,
  my_absLog2FCThreshold = absLg2FCThreshold,
  geneNamesInsteadOfAccessions = FALSE,
  isUniprotDB = TRUE
)

generateTxtFilesForDownstreamAnalysisAndReturnCandidateMatrix_2019(
  comparisonName = "TestComparison",
  mergedResultMatrix,
  my_pModThreshold = pModThreshold,
  my_absLog2FCThreshold = absLg2FCThreshold,
  geneNamesInsteadOfAccessions = FALSE,
  isUniprotDB = TRUE
)
}
\arguments{
\item{comparisonName}{Name for comparison used for the filenames of txt files}

\item{mergedResultMatrix}{result matrix with statistics from 2grp analysis and phosphoSTY information}

\item{my_pModThreshold}{threshold for moderated p.value}

\item{my_absLog2FCThreshold}{threshold absolute log2 fold-change}

\item{geneNamesInsteadOfAccessions}{boolean to inidicate parsing uniprot accession or gene names sp|accession|gene_name}

\item{isUniprotDB}{boolean to indicate if identifier is of type uniprot}
}
\description{
Write txt files with unique protein names that can be used for ORA analysis

Write txt files with unique protein names that can be used for ORA analysis
}
