% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_dataset_docu.R
\docType{data}
\name{genericQuantMatrixGRP2}
\alias{genericQuantMatrixGRP2}
\title{Grp2Analysis reference class for running 2 grp analysis}
\format{reference class}
\usage{
genericQuantMatrixGRP2
}
\description{
Grp2Analysis reference class for running 2 grp analysis
}
\keyword{datasets}
