% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhosphoService_HelperFunctions_v4_2019.R
\name{get2GroupAnalysisFromExistingAnnotationFileNoRender}
\alias{get2GroupAnalysisFromExistingAnnotationFileNoRender}
\title{This function returns the result of the 2grp analysis as tables as well as object}
\usage{
get2GroupAnalysisFromExistingAnnotationFileNoRender(proteinMat,
  pathNFileForAnnotation, mxNAs, nrPep, DenomReference,
  expName = "GlobalProtein_comp1",
  resultDirectory = "myResultDirectory", lg2FCthreshold = qfoldchange,
  qvThreshold = qvalueThreshold, Pairing = FALSE)
}
\arguments{
\item{proteinMat}{quantitative matrix (can alos be peptide matrix)}

\item{pathNFileForAnnotation}{working directory where the annotation file is found}

\item{mxNAs}{maximum number of NAs allowed before rows are filtered-out}

\item{nrPep}{minimum number of peptides}

\item{DenomReference}{reference that is used as denominator in cases where ratios (or logratios) are used (e.g. volcano plot)}

\item{expName}{name for the experiment}

\item{resultDirectory}{directory generated on the fly where the annotation table is saved}

\item{lg2FCthreshold}{threshold for the absolute log2 fold-change (symmetric in both directions)}

\item{qvThreshold}{threshold for qValue (adjust.p.value)}

\item{Pairing}{boolean TRUE/FALSE}
}
\value{
list with results of a twoGroup analysis
}
\description{
This function returns the result of the 2grp analysis as tables as well as object
}
