% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhosphoService_HelperFunctions_v4_2019.R
\name{getCandidatesFromMergedMatrix}
\alias{getCandidatesFromMergedMatrix}
\title{Subset merged matrix based on thresholds for moderated p.Value and absolute log2 fold change}
\usage{
getCandidatesFromMergedMatrix(mergedResultMatrix, my_pModThreshold,
  my_absLog2FCThreshold)
}
\arguments{
\item{mergedResultMatrix}{result matrix with statistics from 2grp analysis and phosphoSTY information}

\item{my_pModThreshold}{threshold for moderated p.value}

\item{my_absLog2FCThreshold}{threshold absolute log2 fold-change}
}
\value{
candidates (subset matching thresholds)
}
\description{
Subset merged matrix based on thresholds for moderated p.Value and absolute log2 fold change
}
