% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhosphoService_HelperFunctions_v4_2019.R
\name{getDFwithMergedQuantResultsWithLargeMatrix}
\alias{getDFwithMergedQuantResultsWithLargeMatrix}
\title{merge statistics from 2grp with the additional information from the phosphoSTY file into one}
\usage{
getDFwithMergedQuantResultsWithLargeMatrix(theGoodMatrix,
  phosPepsQuantResults)
}
\arguments{
\item{phosPepsQuantResults}{the table from the 2grp analysis with the statistics}

\item{myGoodMatrix}{the quant matrix prepared with the prepareSmallerMatrixFromPhospohSTY function}
}
\value{
merged data frame
}
\description{
merge statistics from 2grp with the additional information from the phosphoSTY file into one
}
\examples{
getDFwithMergedQuantResultsWithLargeMatrix(theGoodMatrix = SRMService::myMat_good, phosPepsQuantResults = SRMService::quantResPhosphoPep)

}
