% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limmaHelpers.R
\name{getPVals}
\alias{getPVals}
\title{get topTables for all coefficients from a limma lmfit ebayes result an merge them into single tibble.}
\usage{
getPVals(lmfitebayes, var = "ProteinID")
}
\arguments{
\item{lmfitebayes}{returned by lmFit}

\item{var}{since we work with tibbles what should be the colnames for row.names}
}
\description{
get topTables for all coefficients from a limma lmfit ebayes result an merge them into single tibble.
}
