% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grp2Analysis.R
\docType{class}
\name{Grp2Analysis-class}
\alias{Grp2Analysis}
\alias{Grp2Analysis-class}
\title{Perform 2 group analysis with visualization}
\description{
Perform 2 group analysis with visualization
}
\section{Methods}{

\describe{
\item{\code{getConditionData(condition)}}{get intensities as matrix for single condition}

\item{\code{getNrNAs()}}{return number of NAs per protein}

\item{\code{setMQProteinGroups(MQProteinGroups)}}{set MQ protein groups table}

\item{\code{setProteins(protein)}}{used to verify proteingroups structure and set members}
}}

