% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_dataset_docu.R
\docType{data}
\name{grp2PullDownExample}
\alias{grp2PullDownExample}
\title{Grp2Analysis on pulldown data example}
\format{reference class}
\usage{
grp2PullDownExample
}
\description{
Grp2Analysis on pulldown data example
}
\keyword{datasets}
