% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyMS_R6_AnalysisConfiguration.R
\name{linePlotHierarchy_configuration}
\alias{linePlotHierarchy_configuration}
\title{extracts the relevant information from the configuration to make the plot.}
\usage{
linePlotHierarchy_configuration(res, proteinName, configuration,
  separate = FALSE)
}
\description{
extracts the relevant information from the configuration to make the plot.
}
