% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyMS_R6_AnalysisConfiguration.R
\name{linePlotHierarchy_QuantLine}
\alias{linePlotHierarchy_QuantLine}
\title{make some nice plot with summary statistics}
\usage{
linePlotHierarchy_QuantLine(p, data, aes_y, configuration)
}
\description{
make some nice plot with summary statistics
}
