% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limmaHelpers.R
\name{mergeLimmEBayesResult}
\alias{mergeLimmEBayesResult}
\title{merges results produced by multigroupEBaysQvalue}
\usage{
mergeLimmEBayesResult(multigrpBayes)
}
\arguments{
\item{multigrpBayes}{result of multigroupEBaysQvalue function}
}
\value{
data frame
}
\description{
merges results produced by multigroupEBaysQvalue
}
