% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyMS_R6_AnalysisConfiguration.R
\name{missignessHistogram}
\alias{missignessHistogram}
\title{Histogram summarizing missigness}
\usage{
missignessHistogram(x, configuration, showempty = TRUE, nrfactors = 1)
}
\description{
Histogram summarizing missigness
}
\examples{
missignessHistogram(sample_analysis,skylineconfig)
setNa <- function(x){ifelse(x < 100, NA, x)}
sample_analysis \%>\% mutate(Area = setNa(Area)) -> sample_analysis
missignessHistogram(sample_analysis,skylineconfig)
}
