% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQProteinGroups.R
\name{MQProteinGroups}
\alias{MQProteinGroups}
\title{extract intensities and annotations from MQ proteinGroups.txt}
\usage{
MQProteinGroups(MQProteinGroups)
}
\description{
extract intensities and annotations from MQ proteinGroups.txt
}
