% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_dataset_docu.R
\docType{data}
\name{mqQuantMatrixGRP2}
\alias{mqQuantMatrixGRP2}
\title{Grp2Analysis on of nockout with missing data}
\format{
reference class
}
\usage{
mqQuantMatrixGRP2
}
\description{
Grp2Analysis on of nockout with missing data
}
\keyword{datasets}
