% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMaxQuant.R
\name{MQtoMSstatsFormat}
\alias{MQtoMSstatsFormat}
\title{MaxQtoMSstatsFormat}
\usage{
MQtoMSstatsFormat(evidence, annotation, proteinGroups,
  useUniquePeptide = TRUE, summaryforMultipleRows = max,
  fewMeasurements = "remove", removeMpeptides = TRUE)
}
\arguments{
\item{evidence}{MQ evidence.txt read using read.csv(sep=\"\\t\")}

\item{annotation}{data.frame with columns Raw.file, Condition, BioReplicate, Run, IsotopeLabelType}

\item{proteinGroups}{provide MQ proteinGroups.txt}

\item{useUniquePeptide}{: (likely same as proteotypic) remove peptides that are assigned for more than one proteins. We assume to use unique peptide for each protein.}

\item{summaryforMultipleRows}{: max or sum - when there are multiple measurements for certain feature and certain fun, use highest or sum of all.}

\item{fewMeasurements}{: if 1 or 2 measurements across runs per feature, 'remove' will remove those featuares. It can affected for unequal variance analysis.}
}
\description{
MaxQtoMSstatsFormat
}

