% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limmaHelpers.R
\name{multigroupEBaysQvalue}
\alias{multigroupEBaysQvalue}
\title{runs limma and computes qvalues}
\usage{
multigroupEBaysQvalue(intMatrix, designMatrix)
}
\arguments{
\item{intMatrix}{intensity matrix}

\item{designMatrix}{design matrix}
}
\value{
list with results
}
\description{
runs limma and computes qvalues
}
