% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRMService.R
\docType{class}
\name{PeptideTable-class}
\alias{PeptideTable-class}
\alias{PeptideTable}
\title{Peptide table}
\description{
Peptide table
}
\section{Methods}{

\describe{
\item{\code{removeDecorrelated(minCorrelation = 0.8)}}{removes decorrelated peptides}
}}

