% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyMS_R6_AnalysisConfiguration.R
\name{plot_cv_distribution_violin}
\alias{plot_cv_distribution_violin}
\title{applys func - a funciton workin on matrix for each protein and returning a vector of the same length as the number of samples}
\usage{
plot_cv_distribution_violin(data, config, stat = c("CV", "mean", "sd"))
}
\description{
applys func - a funciton workin on matrix for each protein and returning a vector of the same length as the number of samples
}
\examples{
library(SRMService)
library(tidyverse)
data <- sample_analysis
config <- skylineconfig$clone(deep=TRUE)
res <- summarize_cv(data, config)
plot_cv_distribution_violin(res, config, stat="mean")
plot_cv_distribution_violin(res, config, stat="sd")
plot_cv_distribution_violin(res, config, stat="CV")
}
