% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyMS_R6_TransitionCorrelations.R
\name{plot_intensity_distribution_violin}
\alias{plot_intensity_distribution_violin}
\title{visualize intensity distributions}
\usage{
plot_intensity_distribution_violin(data, config)
}
\description{
visualize intensity distributions
}
\examples{
config <- skylineconfig$clone(deep=TRUE)
analysis <- transformIntensities(sample_analysis, config, log2)
plot_intensity_distribution_violin(analysis, config)
}
