% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{plotNicely}
\alias{plotNicely}
\title{Make nice plots of transitions or peptides with correlations}
\usage{
plotNicely(dataX, main = "")
}
\arguments{
\item{dataX}{data.frame}

\item{main}{some name to plot}
}
\description{
Make nice plots of transitions or peptides with correlations
}

