% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_AnalysisConfiguration.R
\name{plotPeptides}
\alias{plotPeptides}
\title{Plot peptide and fragments}
\usage{
plotPeptides(data, proteinName, sample, intensity, peptide, fragment, factor,
  isotopeLabel, separate = FALSE, log_y = FALSE)
}
\description{
Plot peptide and fragments
}
