% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhosphoService_HelperFunctions_v4_2019.R,
%   R/PhosphoService_HelperFunctions_v5_2020.R
\name{prepareGenericIntensityMatrixFromGoodMatrixForTesting}
\alias{prepareGenericIntensityMatrixFromGoodMatrixForTesting}
\title{prepare generic intensityMatrix from the subsetted and handled matrix before}
\usage{
prepareGenericIntensityMatrixFromGoodMatrixForTesting(myGoodMatrix)

prepareGenericIntensityMatrixFromGoodMatrixForTesting(myGoodMatrix)
}
\arguments{
\item{myGoodMatrix}{the quant matrix prepared with the prepareSmallerMatrixFromPhospohSTY function (note, intensities are log2 in the input and linear in the output)}
}
\value{
data.frame with less columns and rows

data.frame with less columns and rows
}
\description{
prepare generic intensityMatrix from the subsetted and handled matrix before

prepare generic intensityMatrix from the subsetted and handled matrix before
}
\examples{
prepareGenericIntensityMatrixFromGoodMatrixForTesting(SRMService::myMat_good)

prepareGenericIntensityMatrixFromGoodMatrixForTesting(SRMService::myMat_good)

}
