% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhosphoService_HelperFunctions_v5_2020.R
\name{prepareGenericIntensityMatrixFromGoodMatrixForTesting_TMT}
\alias{prepareGenericIntensityMatrixFromGoodMatrixForTesting_TMT}
\title{prepare generic intensityMatrix from the subsetted and handled matrix before}
\usage{
prepareGenericIntensityMatrixFromGoodMatrixForTesting_TMT(myGoodMatrix)
}
\arguments{
\item{myGoodMatrix}{the quant matrix prepared with the prepareSmallerMatrixFromPhospohSTY function (note, intensities are log2 in the input and linear in the output)}
}
\value{
data.frame with less columns and rows
}
\description{
prepare generic intensityMatrix from the subsetted and handled matrix before
}
\examples{
prepareGenericIntensityMatrixFromGoodMatrixForTesting_TMT(SRMService::myMat_good)

}
