% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhosphoService_HelperFunctions_v4_2019.R
\name{prepareSmallerMatrixFromPhospohSTY}
\alias{prepareSmallerMatrixFromPhospohSTY}
\title{prepare smaller data.frame from original MaxQuant PhosphoSTYSite file and reduce to columns that are useful and filter with Localizations probability}
\usage{
prepareSmallerMatrixFromPhospohSTY(phosphoSTYtableFromMQ,
  LocalizationProbabilityFilter = 0.75, min_nonNAinTotal = 3)
}
\arguments{
\item{LocalizationProbabilityFilter}{Threshold on Localization.prob}

\item{min_nonNAinTotal}{minimum number of valid values (non NAs)}

\item{phosphoSTYSitetableFromMQ}{original MaxQuant Phospho_STY_file}
}
\value{
data.frame with less columns and rows
}
\description{
prepare smaller data.frame from original MaxQuant PhosphoSTYSite file and reduce to columns that are useful and filter with Localizations probability
}
