% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhosphoService_HelperFunctions_v5_2020.R
\name{prepareSmallerMatrixFromPhospohSTY_TMT}
\alias{prepareSmallerMatrixFromPhospohSTY_TMT}
\title{prepare smaller data.frame from original MaxQuant PhosphoSTYSite file and reduce to columns that are useful and filter with Localizations probability}
\usage{
prepareSmallerMatrixFromPhospohSTY_TMT(
  phosphoSTYtableFromMQ,
  grepPatternForColumnsOfInterest = "corrected\\\\.\\\\d*___",
  LocalizationProbabilityFilter = 0.75,
  min_nonNAinTotal = 3
)
}
\arguments{
\item{grepPatternForColumnsOfInterest}{pattern to grep over columns of interest}

\item{LocalizationProbabilityFilter}{Threshold on Localization.prob}

\item{min_nonNAinTotal}{minimum number of valid values (non NAs)#'}

\item{phosphoSTYSitetableFromMQ}{original MaxQuant Phospho_STY_file}
}
\value{
data.frame with less columns and rows
}
\description{
prepare smaller data.frame from original MaxQuant PhosphoSTYSite file and reduce to columns that are useful and filter with Localizations probability
}
