% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RequiredColumns.R
\docType{data}
\name{proteinColumns}
\alias{proteinColumns}
\title{Protein Columns}
\format{An object of class \code{character} of length 4.}
\usage{
proteinColumns
}
\description{
Protein Columns
}
\keyword{datasets}
