% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRMService.R
\docType{class}
\name{ProteinTable-class}
\alias{ProteinTable}
\alias{ProteinTable-class}
\title{Protein Table}
\description{
Protein Table
}
\section{Fields}{

\describe{
\item{\code{data}}{data.frame with colnames sample ID rownames proteinID}

\item{\code{conditionmapping}}{data.frame with 2 columns providing mapping of sampleID to condition}

\item{\code{experimentID}}{name of the experiment}
}}

