% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_ProteinTable.R
\docType{class}
\name{ProteinTableR6}
\alias{ProteinTableR6}
\title{Protein Table}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ProteinTableR6
}
\description{
Protein Table
}
\section{Fields}{

\describe{
\item{\code{data}}{data.frame with colnames sample ID rownames proteinID}

\item{\code{experimentID}}{name of the experiment}
}}

\examples{
library(SRMService)
x <- data.frame(Raw.file = c("A","B","C"),"MeasurementOrder" = c(1,2,3) , Gender = c("F","F","M"))
cm <- Annotation$new(x, experimentID = "E1 fixed Gender", fixed = "Gender")
dataProt <- data.frame("proteinID" = c("P1","P2","P3"), NrPeptides = c(1,2,3), "Fasta.Headers" = rep("",3), "A" = rexp(3), "B" = rexp(3), "C" = rexp(3))
length(intersect("Annotation",class(dataProt))) > 0
cm$sampleID
pt <- ProteinTableR6$new(cm)
pt$set_data_wide(dataProt)
pt$get_data_long()
pt$get_data_wide()
pt$get_all_long()
}
\keyword{datasets}
