% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_ProteinVariableSelect.R
\name{ProteinVariableSelect}
\alias{ProteinVariableSelect}
\title{Wraps 4 column matrix, with Condition, Run, Protein, Intensity}
\description{
Wraps 4 column matrix, with Condition, Run, Protein, Intensity

Wraps 4 column matrix, with Condition, Run, Protein, Intensity
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ProteinVariableSelect$new()}}
\item \href{#method-getIntensities}{\code{ProteinVariableSelect$getIntensities()}}
\item \href{#method-getWideFormat}{\code{ProteinVariableSelect$getWideFormat()}}
\item \href{#method-getWideNoMissing}{\code{ProteinVariableSelect$getWideNoMissing()}}
\item \href{#method-getIntensitiesNoMissing}{\code{ProteinVariableSelect$getIntensitiesNoMissing()}}
\item \href{#method-clone}{\code{ProteinVariableSelect$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProteinVariableSelect$new(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getIntensities"></a>}}
\if{latex}{\out{\hypertarget{method-getIntensities}{}}}
\subsection{Method \code{getIntensities()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProteinVariableSelect$getIntensities()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getWideFormat"></a>}}
\if{latex}{\out{\hypertarget{method-getWideFormat}{}}}
\subsection{Method \code{getWideFormat()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProteinVariableSelect$getWideFormat()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getWideNoMissing"></a>}}
\if{latex}{\out{\hypertarget{method-getWideNoMissing}{}}}
\subsection{Method \code{getWideNoMissing()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProteinVariableSelect$getWideNoMissing()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getIntensitiesNoMissing"></a>}}
\if{latex}{\out{\hypertarget{method-getIntensitiesNoMissing}{}}}
\subsection{Method \code{getIntensitiesNoMissing()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProteinVariableSelect$getIntensitiesNoMissing()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProteinVariableSelect$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
