% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCReport.R
\docType{class}
\name{QCProteinReport-class}
\alias{QCProteinReport-class}
\alias{QCProteinReport}
\title{Perform 2 group analysis with visualization}
\description{
Perform 2 group analysis with visualization
}
\section{Fields}{

\describe{
\item{\code{proteinIntensity}}{data.frame where colnames are Raw.File names, row.names are protein ID's and cells are protein abundances.}

\item{\code{proteinAnnotation}}{information about the proteins, nr of peptides etc.}

\item{\code{nrPeptides}}{min number of peptides per protein}

\item{\code{maxNA}}{maximum number of NA's}

\item{\code{projectName}}{name of project}

\item{\code{projectID}}{name of experiment}
}}

\section{Methods}{

\describe{
\item{\code{getNrNAs()}}{return number of NAs per protein}

\item{\code{setMQProteinGroups(MQProteinGroups, debug = FALSE)}}{set MQ protein groups table}

\item{\code{setProteins(protein)}}{used to verify proteingroups structure and set members}
}}

