% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyMS_R6_TransitionCorrelations.R
\name{rankPrecursorsByIntensity}
\alias{rankPrecursorsByIntensity}
\title{ranks precursor - peptide by intensity.}
\usage{
rankPrecursorsByIntensity(data, config)
}
\description{
ranks precursor - peptide by intensity.
}
\section{TODO}{
NA
}

