% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyMS_R6_TransitionCorrelations.R
\name{rankPrecursorsByNAs}
\alias{rankPrecursorsByNAs}
\title{Ranks precursors by NAs (adds new column .NARank)}
\usage{
rankPrecursorsByNAs(data, config)
}
\description{
Ranks precursors by NAs (adds new column .NARank)
}
