% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDProteins.R
\name{remap_PD_ProteinTable}
\alias{remap_PD_ProteinTable}
\title{remap pd proteins to format compatible with grp2Analysis reference class.
can be then set using set using the setProteins method}
\usage{
remap_PD_ProteinTable(proteins, inputFiles)
}
\arguments{
\item{proteins}{read _Proteins.txt file using readr::read_tsv and pass it to the proteins parameter}

\item{inputFiles}{read _InputFiles.txt file using the readr::read_tsv and pass it to inputFiles}
}
\description{
remap pd proteins to format compatible with grp2Analysis reference class.
can be then set using set using the setProteins method
}
