% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareRmdAnalysisEnv.R
\name{RMD_LibraryGen_specLProzor}
\alias{RMD_LibraryGen_specLProzor}
\title{Copies the RMD and Run R file for Library generation your working directory.}
\usage{
RMD_LibraryGen_specLProzor(workdir = getwd())
}
\arguments{
\item{workdir}{specify working directory}
}
\description{
Please see the Run_specLProzor.R file in the working directory
for more details
}
