% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareRmdAnalysisEnv.R
\name{RMD_MQ_Quant_2GrpAnalysis}
\alias{RMD_MQ_Quant_2GrpAnalysis}
\title{copies the RMD and Run files for 2 grp analysis based on MQ (MaxQuant) in your working directory
- ideally the txt directory of your MQ session.}
\usage{
RMD_MQ_Quant_2GrpAnalysis(workdir = getwd())
}
\arguments{
\item{workdir}{specify working directory}
}
\description{
Please see the Run_QuantTwoGroupAnalysis.R for more details
}
\examples{
SRMService::RMD_MQ_Quant_2GrpAnalysis()
}
