% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareRmdAnalysisEnv.R
\name{RMD_PD_Quant_2GrpAnalysis}
\alias{RMD_PD_Quant_2GrpAnalysis}
\title{copies the RMD and Run files for 2 grp analysis based on a PD (protein discoverer) in your working directory.}
\usage{
RMD_PD_Quant_2GrpAnalysis(workdir = getwd())
}
\arguments{
\item{workdir}{specify working directory}
}
\description{
Please see the Run_QuantTwoGroupAnalysis.R for more details
}
\examples{
SRMService::RMD_Quant_2GrpAnalysis()
}
