% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareRmdAnalysisEnv.R
\name{RMD_PhosphoService_FullAnalysis}
\alias{RMD_PhosphoService_FullAnalysis}
\title{Copies the Rnw file and Run R file for Phospho-Service Analysis into your working directory.}
\usage{
RMD_PhosphoService_FullAnalysis(workdir = getwd(), minimal = FALSE)
}
\arguments{
\item{workdir}{specify working directory}
}
\description{
Please see the Run_PhosphoServiceAnalysis.R file in your working directory,
for more details
}
