% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareRmdAnalysisEnv.R
\name{RMD_QC1To1_Old}
\alias{RMD_QC1To1_Old}
\title{Copies the Rnw file and Run R file for old 1To1 QC into your working directory.}
\usage{
RMD_QC1To1_Old(workdir = getwd(), minimal = TRUE)
}
\arguments{
\item{workdir}{specify working directory}
}
\description{
Please see the Run_1To1_oldStyle.R file in your working directory,
for more details
}
