% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareRmdAnalysisEnv.R
\name{RMD_SRM_QC_Normalization}
\alias{RMD_SRM_QC_Normalization}
\title{Copies the Rnw file and Run R file for the SRM QC and normalization.}
\usage{
RMD_SRM_QC_Normalization(workdir = getwd())
}
\arguments{
\item{workdir}{specify working directory}
}
\description{
Please see the Run_1To1_oldStyle.R file in your working directory,
for more details
}
