% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareRmdAnalysisEnv.R
\name{RMD_VarSelection}
\alias{RMD_VarSelection}
\title{Copies the RMD and Run files for variable selection into your working directory.}
\usage{
RMD_VarSelection(workdir = getwd())
}
\description{
Please see the Run_QuantTwoGroupAnalysis.R for more details
}
