% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareRmdAnalysisEnv.R
\name{RMD_VarSelection_QCReport}
\alias{RMD_VarSelection_QCReport}
\title{Copies the RMD and Run files for 2 grp analysis in your working directory.
Please see the Run_QuantTwoGroupAnalysis.R for more details}
\usage{
RMD_VarSelection_QCReport()
}
\description{
Copies the RMD and Run files for 2 grp analysis in your working directory.
Please see the Run_QuantTwoGroupAnalysis.R for more details
}
\examples{
SRMService::RMD_VarSelection_QCReport()

}

