% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_dataset_docu.R
\docType{data}
\name{sample_analysis}
\alias{sample_analysis}
\title{A data frame wich goes along with the \link{skylineconfig}.}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{protein_Id}{protein id}
  \item{peptide_Id}{peptide id - stripped sequence}
  ...
}}
\source{
\url{http://www.fgcz.ch/}
}
\usage{
sample_analysis
}
\description{
A data frame wich goes along with the \link{skylineconfig}.
}
\keyword{datasets}
