% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyMS_R6_TransitionCorrelations.R
\name{setIntensitiesToNA}
\alias{setIntensitiesToNA}
\title{sets intensities to NA if maxQValue_Threshold exceeded}
\usage{
setIntensitiesToNA(data, config, newcolname = "IntensitiesWithNA")
}
\value{
list with augmented data and updated config
}
\description{
sets intensities to NA if maxQValue_Threshold exceeded
}
