% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyMS_R6_TransitionCorrelations.R
\name{setLarge_Q_ValuesToNA}
\alias{setLarge_Q_ValuesToNA}
\title{sets intensities to NA if maxQValue_Threshold exceeded}
\usage{
setLarge_Q_ValuesToNA(data, config, intensityNewName = "IntensitiesWithNA")
}
\description{
sets intensities to NA if maxQValue_Threshold exceeded
}
\examples{
analysis <- SRMService::spectronautDIAData250_analysis
config <- SRMService::spectronautDIAData250_config$clone(deep=TRUE)
res <- setLarge_Q_ValuesToNA(analysis, config)
}
