% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyMS_R6_TransitionCorrelations.R
\name{setSmallIntensitiesToNA}
\alias{setSmallIntensitiesToNA}
\title{sets intensities smaller than threshold to NA}
\usage{
setSmallIntensitiesToNA(data, config, threshold = 1,
  intensityNewName = "IntensitiesWithNA")
}
\description{
sets intensities smaller than threshold to NA
}
\examples{

analysis <- SRMService::spectronautDIAData250_analysis
config <- SRMService::spectronautDIAData250_config$clone(deep=TRUE)

config$table$getWorkIntensity()

config2 <- config$clone(deep=TRUE)
res1 <- setSmallIntensitiesToNA(analysis, config, threshold=1, intensityNewName = "Int1" )
res1000 <- setSmallIntensitiesToNA(analysis, config2, threshold=1000, intensityNewName = "Int1000" )
sum(is.na(res1[[config$table$getWorkIntensity()]])) < sum(is.na(res1000[[config2$table$getWorkIntensity()]]))
}
\seealso{
Other filter functions: \code{\link{removeLarge_Q_Values}}
}
