% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyMS_R6_AnalysisConfiguration.R
\name{setup_analysis}
\alias{setup_analysis}
\title{Extracts columns relevant for a configuration from a data frame}
\usage{
setup_analysis(data, configuration, sep = "~")
}
\description{
Extracts columns relevant for a configuration from a data frame
}
\examples{

skylineconfig <- craeteSkylineConfiguration(isotopeLabel="Isotope.Label.Type", qValue="Detection.Q.Value")
skylineconfig$table$factors[["Time"]] = "Sampling.Time.Point"
data(skylinePRMSampleData)

sample_config <- setup_analysis(skylinePRMSampleData, skylineconfig)

}
