% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_dataset_docu.R
\docType{data}
\name{skylineconfig}
\alias{skylineconfig}
\title{A configuration which matches the \link{sample_analysis} data.}
\format{A AnalysisConfiguration R6 class}
\usage{
skylineconfig
}
\description{
A configuration which matches the \link{sample_analysis} data.
}
\keyword{datasets}
