% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRMService.R
\docType{class}
\name{SRMService-class}
\alias{SRMService}
\alias{SRMService-class}
\title{R access to Bibliospec File}
\description{
This class implements an R referenz class for BiblioSpec
generated sqlite files and can return the data contained as data.frames
or as list of tandem mass spectra peptide assignments objects (psm).
}
\details{
The function performs a SQL query on the SQLite
}
\section{Fields}{

\describe{
\item{\code{dbfile}}{database file location}
}}
\examples{
tmp <- "D:/googledrive/DataAnalysis/p1930"
allData <- read.csv( file=file.path(tmp,"data/longFormat.txt"),row.names = 1)
data <-allData
pool=1
data <-allData[allData$pool==pool,]
srms <- SRMService(data)

}

