% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyMS_R6_TransitionCorrelations.R
\name{summariseNAs}
\alias{summariseNAs}
\title{add number of NA's at lowest hierarchy to data}
\usage{
summariseNAs(data, config)
}
\description{
add number of NA's at lowest hierarchy to data
}
\examples{

config <- spectronautDIAData250_config$clone(deep=T)
res <- setLarge_Q_ValuesToNA(spectronautDIAData250_analysis, config)
res <- summariseNAs(res,config)
x <- res \%>\%
  select(config$table$hierarchyKeys()[1], config$table$hierarchyKeys(T)[1], "srm_NrNAs") \%>\%
  distinct() \%>\% summarize(sum(srm_NrNAs)) \%>\% pull()
stopifnot(sum(is.na(res[[config$table$getWorkIntensity()[1]]])) == x)
}
