% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TidyMS.R, R/tidyMS_R6_TransitionCorrelations.R
\name{summariseQValues}
\alias{summariseQValues}
\alias{summariseQValues}
\title{Compute QValue summaries for each precursor}
\usage{
summariseQValues(data, precursorId = getConfig(data)$.PrecursorId,
  QValue = config_col_map(data)$QValue,
  maxQValThreshold = config_parameters(data)$maxQValue_Threshold,
  minNumberOfQValues = config_parameters(data)$minNumberOfQValues)

summariseQValues(data, precursorId = getConfig(data)$.PrecursorId,
  QValue = config_col_map(data)$QValue,
  maxQValThreshold = config_parameters(data)$maxQValue_Threshold,
  minNumberOfQValues = config_parameters(data)$minNumberOfQValues)
}
\arguments{
\item{data}{1}

\item{precursorId}{2}

\item{QValue}{3}

\item{maxQValThreshold}{4}

\item{minNumberOfQValues}{5}

\item{data}{1}

\item{precursorId}{2}

\item{QValue}{3}

\item{maxQValThreshold}{4}

\item{minNumberOfQValues}{5}
}
\description{
Compute QValue summaries for each precursor

Compute QValue summaries for each precursor
}
