% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyMS_R6_AnalysisConfiguration.R
\name{summarizeHierarchy}
\alias{summarizeHierarchy}
\title{Summarize peptide Counts}
\usage{
summarizeHierarchy(x, configuration, hierarchy_level = 1, factor_level = 0)
}
\description{
Summarize peptide Counts
}
\examples{
library(SRMService)
library(tidyverse)
<<<<<<< HEAD
skylineconfig <- craeteSkylineConfiguration(isotopeLabel="Isotope.Label.Type", ident_qValue="Detection.Q.Value")
=======
skylineconfig <- createSkylineConfiguration(isotopeLabel="Isotope.Label.Type", ident_qValue="Detection.Q.Value")
>>>>>>> d5382ac60d502c8d041c3cc4433564e7eaa8e6f6
skylineconfig$table$factors[["Time"]] = "Sampling.Time.Point"
data(skylinePRMSampleData)

sample_analysis <- setup_analysis(skylinePRMSampleData, skylineconfig)
res <- summarizeHierarchy(sample_analysis, skylineconfig)
res2 <- summarizeHierarchy(sample_analysis, skylineconfig, factor_level=1)
summarizeHierarchy(sample_analysis, skylineconfig, hierarchy_level =2, factor_level=0 )
summarizeHierarchy(sample_analysis, skylineconfig, hierarchy_level =3 )
summarizeHierarchy(sample_analysis, skylineconfig, hierarchy_level =4 )

}
