% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQProteinGroups.R
\name{tidyMQ_Peptides}
\alias{tidyMQ_Peptides}
\title{parse MQ peptides.txt}
\usage{
tidyMQ_Peptides(MQPeptides)
}
\arguments{
\item{MQPeptides}{data.frame generated with read.csv("peptide.txt",sep="\\t", stringsAsFactors=FALSE)}
}
\description{
parse MQ peptides.txt
}
\examples{
library(tidyverse)
peptides_txt <- system.file("samples/maxquant_txt/MSQC1/peptides.txt",package = "SRMService")
peptides_txt <- read.csv(peptides_txt, header=TRUE, stringsAsFactors = FALSE, sep="\\t")
str(peptides_txt)
tmp <-paste(peptides_txt$Evidence.IDs, collapse = ";")
tmp <- strsplit(tmp, ";")
length(unique(tmp[[1]]))
res <-tidyMQ_Peptides(peptides_txt)
dim(res)
head(res)
}
