% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQProteinGroups.R
\name{tidyMQ_ProteinGroups}
\alias{tidyMQ_ProteinGroups}
\title{extract intensities and annotations from MQ proteinGroups.txt}
\usage{
tidyMQ_ProteinGroups(MQProteinGroups)
}
\arguments{
\item{MQProteinGroups}{data.frame generated with read.csv("peptide.txt",sep="\\t", stringsAsFactors=FALSE)}
}
\description{
extract intensities and annotations from MQ proteinGroups.txt
}
\examples{
protein_txt <- system.file("samples/maxquant_txt/MSQC1/proteinGroups.txt",package = "SRMService")
protein_txt
protein_txt <- read.csv(protein_txt, header=TRUE, stringsAsFactors = FALSE, sep="\\t")
res <-tidyMQ_ProteinGroups(protein_txt)
head(res)
}
