% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyMS_R6_TransitionCorrelations.R
\name{transformIntensities}
\alias{transformIntensities}
\title{Transform intensity}
\usage{
transformIntensities(data, config, transformation, intesityNewName = NULL)
}
\description{
Transform intensity
}
\examples{
library(tidyverse)
config <- SRMService::spectronautDIAData250_config$clone(deep=TRUE)
analysis <- SRMService::spectronautDIAData250_analysis
x <- transformIntensities(analysis, config, transform = log2)
stopifnot("log2_FG.Quantity" \%in\% colnames(x))
config <- SRMService::spectronautDIAData250_config$clone(deep=TRUE)
analysis <- SRMService::spectronautDIAData250_analysis
x <- transformIntensities(analysis, config, transform = asinh)
stopifnot("asinh_FG.Quantity" \%in\% colnames(x))
}
