% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyMS_R6_workflowFunctions.R
\name{workflow_correlation_preprocessing}
\alias{workflow_correlation_preprocessing}
\title{correlation preprocessing}
\usage{
workflow_correlation_preprocessing(data, config)
}
\description{
correlation preprocessing
}
\examples{
config <- spectronautDIAData250_config$clone(deep=T)

data <- spectronautDIAData250_analysis
workflow_correlation_preprocessing(data,config)
}
