% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyMS_R6_TransitionCorrelations.R
\name{aggregateTopNIntensities}
\alias{aggregateTopNIntensities}
\title{aggregates top N intensities}
\usage{
aggregateTopNIntensities(data, config, N = 3)
}
\description{
run \link{rankPrecursorsByIntensity} first
}
