% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProteinTable.R
\docType{class}
\name{Annotation}
\alias{Annotation}
\title{Condition Map}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Annotation
}
\description{
Condition Map
}
\examples{

x <- data.frame(Raw.file = c("A","B","C"),"MeasOrder" = c(1,2,3) , Gender = c("F","F","M"))
cm <- Annotation$new(x, fixed = "Gender")
cm$exists("Gender")
cm$exists("Stop")
cm$fixed
cm$random
cm$as_numeric("Gender")
cm$levels("Gender")
cm$get("Gender")
cm$get_color("Gender")
cm$get_required()
}
\keyword{datasets}
