% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyMS_R6_AnalysisConfiguration.R
\name{extractIntensities}
\alias{extractIntensities}
\title{Extract intensity column}
\usage{
extractIntensities(x, configuration)
}
\description{
Extract intensity column
}
\examples{
rm(skylineconfig)
xx <- extractIntensities(sample_analysis,skylineconfig)
head(xx)
}
