% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grp2Analysis.R
\docType{class}
\name{Grp2Analysis-class}
\alias{Grp2Analysis-class}
\alias{Grp2Analysis}
\title{Perform 2 group analysis with visualization}
\description{
Perform 2 group analysis with visualization
}
\section{Fields}{

\describe{
\item{\code{proteinIntensity}}{data.frame where colnames are Raw.File names, row.names are protein ID's and cells are protein abundances.}

\item{\code{annotation_}}{annotations data.frame with columns such as Raw.File, Condition, Run etc.}

\item{\code{proteinAnnotation}}{information about the proteins, nr of peptides etc.}

\item{\code{nrPeptides}}{min number of peptides per protein}

\item{\code{maxNA}}{maximum number of NA's}

\item{\code{condition}}{summary of conditions}

\item{\code{projectName}}{name of project}

\item{\code{projectID}}{name of experiment}

\item{\code{pvalue}}{pvalue threshold}

\item{\code{qvalue}}{qvalue threshold}

\item{\code{pfoldchange}}{foldchange threshold for p volcano plot}

\item{\code{qfoldchange}}{foldchange threshold for q volcano plot}

\item{\code{reference}}{document}

\item{\code{removeDates}}{strip the date from file name.}
}}

\section{Methods}{

\describe{
\item{\code{getConditionData(condition)}}{get intensities as matrix for single condition}

\item{\code{getModPValuesCI()}}{get Pvalues and confidence intervals}

\item{\code{getNormalized()}}{return normalized data}

\item{\code{getNormalizedGrpAverages()}}{computes grp averages per protein}

\item{\code{getNrNAs()}}{return number of NAs per protein}

\item{\code{getResultTableWithPseudo()}}{add pseudo p-values and pseudo fold changes}

\item{\code{getResultTableWithPseudoAndClustering()}}{calls getResultTableWithPseudo() and adds clusterID column}

\item{\code{setMQProteinGroups(MQProteinGroups)}}{set MQ protein groups table}

\item{\code{setNormalizationMethod(normalizationMethod = "robustscale", housekeeper = "")}}{set the normalization parameters}

\item{\code{setProteins(protein)}}{used to verify proteingroups structure and set members}
}}

