% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_AnalysisConfiguration.R
\name{plotPeptides}
\alias{plotPeptides}
\title{Plot peptide and fragments}
\usage{
plotPeptides(data, proteinName, x, y, group, color, factor, isotopeLabel)
}
\description{
Plot peptide and fragments
}
