% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProteinTable.R
\docType{class}
\name{ProteinTableR6}
\alias{ProteinTableR6}
\title{Protein Table}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ProteinTableR6
}
\description{
Protein Table
}
\section{Fields}{

\describe{
\item{\code{data}}{data.frame with colnames sample ID rownames proteinID}

\item{\code{conditionmapping}}{data.frame with 2 columns providing mapping of sampleID to condition}

\item{\code{experimentID}}{name of the experiment}
}}

\examples{
x <- data.frame(Raw.file = c("A","B","C"),"MeasOrder" = c(1,2,3) , Gender = c("F","F","M"))
cm <- Annotation$new(x, fixed = "Gender")
#pt <- ProteinTableR6$new(cm)
}
\keyword{datasets}
