% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_ProteinVariableSelect.R
\docType{data}
\name{ProteinVariableSelect}
\alias{ProteinVariableSelect}
\title{Wraps 4 column matrix, with Condition, Run, Protein, Intensity}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ProteinVariableSelect
}
\description{
Wraps 4 column matrix, with Condition, Run, Protein, Intensity
}
\keyword{datasets}
