% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareRmdAnalysisEnv.R
\name{RMD_Quant_QCReport}
\alias{RMD_Quant_QCReport}
\title{copies the RMD and Run files for 2 grp analysis in your working directory.
Please see the Run_QuantTwoGroupAnalysis.R for more details}
\usage{
RMD_Quant_QCReport()
}
\description{
copies the RMD and Run files for 2 grp analysis in your working directory.
Please see the Run_QuantTwoGroupAnalysis.R for more details
}
\examples{
SRMService::RMD_QC_QuantReport()

}

