% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyMS_R6_TransitionCorrelations.R
\name{setIntensitiesToNA}
\alias{setIntensitiesToNA}
\title{DEPRECATED use \link{setLarge_Q_ValuesToNA}}
\usage{
setIntensitiesToNA(data, config, newcolname = "IntensitiesWithNA")
}
\description{
DEPRECATED use \link{setLarge_Q_ValuesToNA}
}
