% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRMService.R
\docType{class}
\name{SRMService-class}
\alias{SRMService}
\alias{SRMService-class}
\title{R access to Bibliospec File}
\description{
This class implements an R referenz class for BiblioSpec
generated sqlite files and can return the data contained as data.frames
or as list of tandem mass spectra peptide assignments objects (psm).
}
\details{
The function performs a SQL query on the SQLite
}
\section{Fields}{

\describe{
\item{\code{dbfile}}{database file location}
}}
\section{Methods}{

\describe{
\item{\code{getMatchingIntensities()}}{get matrix with intensities, where nr of NAs in row < maxNA}

\item{\code{getNrNAs(light = FALSE)}}{show nr of NA's for heavy (defalt) or light transitions}

\item{\code{getTransitionIntensities(maxNA, light = FALSE)}}{get matrix with intensities, where nr of NAs in row < maxNA}

\item{\code{maxNAHeavy(max)}}{set maximum of na's heavy row}

\item{\code{maxNALight(max)}}{set maximum of na's in light row}

\item{\code{plotCommonTransitions(light = FALSE)}}{Shows transitions which occure in heavy and light}

\item{\code{plotQValues()}}{show q value distribution for peak groups}
}}
\examples{

library(SRMService)
tmp <- "D:/googledrive/DataAnalysis/p1930"
allData <- read.csv( file=file.path(tmp,"data/longFormat.txt"),row.names = 1)
head(allData)
data <-allData
pool=2
data <-allData[allData$pool==pool,]
head(data)
srms <- SRMService(data,qvalue=0.05)
SRMService$methods()
srms$maxNAHeavy()
SRMService$fields()
head(srms$piw)

srms$plotQValues()
srms$plotTransition()
srms$plotTransition(light=TRUE)

srms$getNrNAs()
srms$getNrNAs(light=TRUE)
srms$maxNAHeavy()
srms$maxNALight()
srms$maxNAHeavy(80)
srms$maxNALight(80)
srms$plotCommonTransitions()
srms$plotCommonTransitions(light=TRUE)
tmp <-srms$getLHLog2FoldChange(maxNA = 20)
srms$maxNAFC()
resH <- srms$plotCommonTransitions()
dim(resH)
resL <-srms$plotCommonTransitions(light=TRUE)
dim(resL)
resAll <- srms$getMatchingIntensities()
dim(resAll$light)

tmpH <- srms$getTransitionIntensities()
tmpL <- srms$getTransitionIntensities(light=TRUE)
dim(tmpL)

}

