% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyMS_R6_TransitionCorrelations.R
\name{summariseNAs}
\alias{summariseNAs}
\title{add number of NA's at lowest hierarchy to data}
\usage{
summariseNAs(data, config)
}
\description{
add number of NA's at lowest hierarchy to data
}
