% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TidyMS.R
\name{summariseQValues}
\alias{summariseQValues}
\title{Compute QValue summaries for each precursor}
\usage{
summariseQValues(data, precursorId = getConfig(data)$.PrecursorId,
  QValue = config_col_map(data)$QValue,
  maxQValThreshold = config_parameters(data)$maxQValue_Threshold,
  minNumberOfQValues = config_parameters(data)$minNumberOfQValues)
}
\arguments{
\item{data}{1}

\item{QValue}{3}

\item{.PrecursorId}{2}
}
\description{
Compute QValue summaries for each precursor
}
