% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyMS_R6_AnalysisConfiguration.R
\name{summarizeHierarchy}
\alias{summarizeHierarchy}
\title{Summarize peptide Counts}
\usage{
summarizeHierarchy(x, configuration, level = 1)
}
\description{
Summarize peptide Counts
}
\examples{
library(SRMService)
skylineconfig <- craeteSkylineConfiguration(isotopeLabel="Isotope.Label.Type", qValue="Detection.Q.Value")
skylineconfig$table$factors[["Time"]] = "Sampling.Time.Point"
data(skylinePRMSampleData)

sample_analysis <- setup_analysis(skylinePRMSampleData, skylineconfig)
summarizeHierarchy(sample_analysis, skylineconfig)
summarizeHierarchy(sample_analysis, skylineconfig, level =2 )
summarizeHierarchy(sample_analysis, skylineconfig, level =3 )
summarizeHierarchy(sample_analysis, skylineconfig, level =4 )
}
