% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQProteinGroups.R
\name{tidyMQ_Peptides}
\alias{tidyMQ_Peptides}
\title{parse MQ peptides.txt}
\usage{
tidyMQ_Peptides(MQPeptides)
}
\arguments{
\item{MQPeptides}{data.frame generated with read.csv("peptide.txt",sep="\\t", stringsAsFactors=FALSE)}
}
\description{
parse MQ peptides.txt
}
