% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQProteinGroups.R
\name{tidyMQ_ProteinGroups}
\alias{tidyMQ_ProteinGroups}
\title{extract intensities and annotations from MQ proteinGroups.txt}
\usage{
tidyMQ_ProteinGroups(MQProteinGroups)
}
\description{
extract intensities and annotations from MQ proteinGroups.txt
}
