% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{transitionCorrelationsJack}
\alias{transitionCorrelationsJack}
\title{Compute correlation matrix}
\usage{
transitionCorrelationsJack(dataX)
}
\arguments{
\item{dataX}{data.frame with transition intensities per peptide}
}
\description{
Compute correlation matrix
}
