% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRMService.R
\docType{class}
\name{TransitionTable-class}
\alias{TransitionTable}
\alias{TransitionTable-class}
\title{Transition table}
\description{
Transition table
}
\section{Methods}{

\describe{
\item{\code{rownamesAsTable(x)}}{help function}
}}

