% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRMService.R
\docType{class}
\name{TransitionTable-class}
\alias{TransitionTable-class}
\alias{TransitionTable}
\title{Transition table}
\description{
Transition table
}
\section{Methods}{

\describe{
\item{\code{getLongFormat()}}{return data in long format (can easily be converted to msstats input)}

\item{\code{getPeptidesAsList()}}{returns list of matrices each matrix representing single peptide}

\item{\code{getProteinsAsList()}}{Returns the data as an list of dataframes}

\item{\code{removeDecorrelated(minCorrelation = 0.8)}}{removes decorrelated peptides}

\item{\code{rownamesAsTable(x)}}{help function}
}}

