% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyMS_R6_workflowFunctions.R
\name{workflow_NA_preprocessing}
\alias{workflow_NA_preprocessing}
\title{filter by NA's and factor information}
\usage{
workflow_NA_preprocessing(data, config, percent = 60)
}
\description{
filter by NA's and factor information
}
\examples{
library(SRMService)
library(tidyverse)
rm(list=ls())
config <- spectronautDIAData250_config$clone(deep=T)
data <- spectronautDIAData250_analysis
tmp <-workflow_NA_preprocessing(data, config)
stopifnot(isFALSE(is.grouped_df(tmp)))
}
\seealso{
Other workflow functions: \code{\link{workflow_correlation_preprocessing}}
}
